Align chunk size to HuBERT frame boundary (20ms steps)
- Change default chunk_sec from 150ms to 160ms (20ms boundary)
- Add 20ms step rounding to chunk size slider in GUI
- Fix HuBERT loader to use proper dtype instead of always float32

HuBERT processes audio in 20ms frames (320 samples @ 16kHz), so
aligning chunk sizes to this boundary improves feature consistency.

+++ rcwx/models/hubert_loader.py
@@ -252,7 +252,7 @@ class HuBERTLoader:
         if audio.dim() == 1:
             audio = audio.unsqueeze(0)
 
-        audio = audio.to(self.device).float()
+        audio = audio.to(self.device).to(self.dtype)
 
         outputs = self.model(audio, output_hidden_states=True)
         features = outputs.hidden_states[output_layer]
